Restructure mittnitectl commands into own package
