+++ cmd/mittnitectl/job_status.go
@@ -1,4 +1,4 @@
-package cmd
+package main
 
 import (
 	"errors"
