+++ cmd/root.go
@@ -27,9 +27,3 @@ func Execute() {
 		log.Fatal(err)
 	}
 }
-
-func ExecuteCTL() {
-	if err := ctlCommand.Execute(); err != nil {
-		log.Fatal(err)
-	}
-}
